requirements.txt
Django==2.2
djangorestframework
